+++ google-cloud/flatcar-linux/kubernetes/controllers.tf
@@ -42,7 +42,7 @@ resource "google_compute_instance" "controllers" {
     auto_delete = true
 
     initialize_params {
-      image = var.os_image
+      image = data.google_compute_image.flatcar-linux.self_link
       size  = var.disk_size
     }
   }