+++ google-cloud/flatcar-linux/kubernetes/workers/variables.tf
@@ -36,7 +36,13 @@ variable "machine_type" {
 
 variable "os_image" {
   type        = string
-  description = "Flatcar Linux image for compute instanges (e.g. gcloud compute images list)"
+  description = "Flatcar Linux image for compute instances (flatcar-stable, flatcar-beta, flatcar-alpha)"
+  default     = "flatcar-stable"
+
+  validation {
+    condition     = contains(["flatcar-stable", "flatcar-beta", "flatcar-alpha"], var.os_image)
+    error_message = "The os_image must be flatcar-stable, flatcar-beta, or flatcar-alpha."
+  }
 }
 
 variable "disk_size" {